NFDIV-5309 Rename & Relocate makeBailiffDecision event & tests
This is an LA event not a CW event.
